backend/requirements.txt
# ===== CORE FRAMEWORK =====
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ===== DATABASE =====
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3

# ===== AUTHENTICATION & SECURITY =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator>=2.0.0
python-dotenv==1.0.1

# ===== AI & RAG =====
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.3
langchain-chroma==0.1.4
chromadb==0.5.18
sentence-transformers==3.0.1
openai==1.52.2

# ===== CLOUD STORAGE =====
cloudinary==1.41.0

# ===== HTTP & CORS =====
aiofiles==24.1.0
starlette==0.38.5

# ===== TESTING (Development only) =====
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0

# ===== OPTIONAL =====
sendgrid==6.11.0
 sentry-sdk[fastapi]==2.19.0
 gunicorn==23.0.0
